Rename package blog → tickets; update imports & Procfile
