+++ Procfile
@@ -1,1 +1,1 @@
-web: uvicorn blog.main:app --host 0.0.0.0 --port 10000
+uvicorn tickets.main:app --host 0.0.0.0 --port $PORT
