+++ tickets/hashing.py
