+++ tickets/schemas/user.py
@@ -5,7 +5,7 @@ from enum import Enum
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from blog.schemas.ticket import TicketOut
+    from tickets.schemas.ticket import TicketOut
 
 
 class UserRole(str, Enum):
